Time: 0 ms (100%), Space: 7.8 MB (61.42%) - LeetHub

+++ 2579-count-total-number-of-colored-cells/2579-count-total-number-of-colored-cells.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long coloredCells(int n) {
        
-       long long x= n * (n-1);
+       long long x= (long long)n * (n-1);
        x = x/2;
 
         return 1+4*x;
